% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesops_plot_emotion_distribution.R
\name{aesops_plot_emotion_distribution}
\alias{aesops_plot_emotion_distribution}
\title{Plot emotion distribution across Aesop’s Fables}
\usage{
aesops_plot_emotion_distribution()
}
\value{
A ggplot object showing emotion counts.
}
\description{
Visualizes the frequency of NRC emotion categories across the full
Aesop’s Fables corpus.
}
