% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesops_sentiment.R
\name{aesops_sentiment}
\alias{aesops_sentiment}
\title{Join Aesop's Fables tokens with sentiment lexicons}
\usage{
aesops_sentiment(method = c("bing", "afinn", "loughran", "nrc"))
}
\arguments{
\item{method}{Sentiment lexicon to use.}
}
\value{
A tibble of tokens joined with sentiment values.
}
\description{
Join Aesop's Fables tokens with sentiment lexicons
}
